+++ src/keyboard/mod.rs
@@ -447,6 +447,11 @@ pub enum Event<'a> {
         /// delay (in milisecond) between a key press and the start of repetition
         delay: i32,
     },
+    /// The key modifiers have changed state
+    Modifiers {
+        /// current state of the modifiers
+        modifiers: ModifiersState,
+    }
 }
 
 /// An event sent at repeated intervals for certain keys determined by xkb_keymap_key_repeats
@@ -775,6 +780,7 @@ where
                         ..
                     } => {
                         state.update_modifiers(mods_depressed, mods_latched, mods_locked, group);
+                        event_impl(Event::Modifiers { modifiers: state.mods_state }, proxy);
                         if key_held.is_some() {
                             state_chan.lock().unwrap().0.send(()).unwrap();
                         }